@@ -1,18 +1,15 @@
 {
   "name": "v-viewer",
-  "version": "2.0.0-alpha",
-  "private": true,
+  "version": "3.0.0",
   "description": "Image viewer component for vue, supports rotation, scale, zoom and so on, based on viewer.js",
   "main": "dist/v-viewer.umd.min.js",
   "types": "types/index.d.ts",
   "scripts": {
-    "dev": "vue-cli-service serve --mode example",
-    "example": "vue-cli-service build --mode example",
-    "release": "vue-cli-service build --mode release --target lib src/index.js",
+    "serve-example": "vue-cli-service serve --mode example",
+    "build-example": "vue-cli-service build --mode example",
+    "build": "vue-cli-service build --mode release --target lib --name v-viewer ./src/main.ts",
     "lint": "vue-cli-service lint",
-    "lint-example": "vue-cli-service lint example",
-    "test:unit": "vue-cli-service test:unit",
-    "cover": "cross-env NODE_ENV=coverage nyc --reporter=lcov --reporter=text npm run test:unit",
+    "lint fix": "vue-cli-service lint --fix",
     "ui": "vue ui",
     "gh": "node deploy",
     "publish": "npm publish"
@@ -35,40 +32,45 @@
     "url": "https://github.com/mirari/v-viewer/issues"
   },
   "homepage": "https://github.com/mirari/v-viewer#readme",
-  "nyc": {
-    "instrument": false,
-    "sourceMap": false
-  },
   "dependencies": {
-    "gh-pages": "^2.1.1",
-    "throttle-debounce": "^2.1.0",
-    "viewerjs": "^1.3.6"
+    "core-js": "^3.6.5",
+    "lodash": "^4.17.21",
+    "throttle-debounce": "^3.0.1",
+    "viewerjs": "^1.9.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.11.0",
-    "@vue/cli-plugin-eslint": "^3.11.0",
-    "@vue/cli-plugin-pwa": "^3.11.0",
-    "@vue/cli-plugin-unit-mocha": "^3.11.0",
-    "@vue/cli-service": "^3.11.0",
-    "@vue/eslint-config-standard": "^4.0.0",
-    "@vue/test-utils": "1.0.0-beta.29",
-    "babel-eslint": "^10.0.1",
-    "babel-plugin-istanbul": "^5.2.0",
-    "bulma": "^0.7.5",
-    "chai": "^4.1.2",
-    "core-js": "^2.6.5",
-    "cross-env": "^5.2.1",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "github-markdown-css": "^3.0.1",
-    "istanbul-instrumenter-loader": "^3.0.1",
-    "node-sass": "^4.9.0",
-    "nyc": "^14.1.1",
-    "register-service-worker": "^1.6.2",
-    "sass-loader": "^7.1.0",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "vue": "^2.6.10",
-    "vue-markdown-loader": "^2.4.1",
-    "vue-template-compiler": "^2.6.10"
+    "@types/lodash": "^4.14.168",
+    "@typescript-eslint/eslint-plugin": "^4.18.0",
+    "@typescript-eslint/parser": "^4.18.0",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-typescript": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "@vue/eslint-config-standard": "^5.1.2",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "bulma": "^0.9.2",
+    "eslint": "^6.7.2",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-vue": "^7.0.0",
+    "lint-staged": "^9.5.0",
+    "sass": "^1.26.5",
+    "sass-loader": "^8.0.2",
+    "typescript": "~4.1.5"
+  },
+  "peerDependencies": {
+    "vue": "^3.0.0"
+  },
+  "gitHooks": {
+    "pre-commit": "lint-staged"
+  },
+  "lint-staged": {
+    "*.{js,jsx,vue,ts,tsx}": [
+      "vue-cli-service lint",
+      "git add"
+    ]
   }
 }
