@@ -1,7 +1,5 @@
 module.exports = {
   presets: [
-    ['@vue/app', {
-      useBuiltIns: false,
-    }],
+    '@vue/cli-plugin-babel/preset',
   ],
 }
